upload_search/requirements.txt
aiohappyeyeballs==2.4.3
aiohttp==3.10.8
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.6.0
attrs==24.2.0
blinker==1.8.2
certifi==2024.8.30
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
distro==1.9.0
et-xmlfile==1.1.0
Flask==3.0.3
frozenlist==1.4.1
gunicorn==23.0.0
h11==0.14.0
httpcore==1.0.6
httpx==0.27.2
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.4
jiter==0.5.0
lxml==5.3.0
MarkupSafe==2.1.5
multidict==6.1.0
numpy==2.1.1
openai==0.28.0
openpyxl==3.1.5
packaging==24.1
pandas==2.2.3
pandas-stubs==2.2.2.240909
psycopg2==2.9.9
pydantic==2.9.2
pydantic_core==2.23.4
PyMuPDF==1.24.10
PyMuPDFb==1.24.10
python-dateutil==2.9.0.post0
python-docx==1.1.2
pytz==2024.2
requests==2.32.3
six==1.16.0
sniffio==1.3.1
tqdm==4.66.5
types-pytz==2024.2.0.20240913
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
Werkzeug==3.0.4
yarl==1.13.1
PyPDF2==3.0.1

openai[datalib]
